requirements.txt
djangorestframework
dialogflow
spotipy